backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
alembic
pydantic
pydantic-settings
python-multipart
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
email-validator


